Combine filters with sort by.
Changing the filter resets any previously selected sort by option.
This is due to the component not updating if category or status filters
are changed.

A dirty fix would be adding all possible combinations as rt-if's,
however for now I consider resetting the sort by selection as a better
solution.